refactor: Move `plugin_cli` implementation to private module `singer_sdk.cli._decorator`
